moves rpc error conversion traits to their own module
